% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analys.R
\name{analys}
\alias{analys}
\title{Select features based on mirror stats.}
\usage{
analys(mm, ww, q)
}
\arguments{
\item{mm}{mirror statistics}

\item{ww}{absolute value of mirror statistics}

\item{q}{FDR control level}
}
\value{
selected index
}
\description{
Select features based on mirror stats.
}
