% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DS.R
\name{DS}
\alias{DS}
\title{FDR control via DS and MDS in linear regression}
\usage{
DS(X, y, num_split, q)
}
\arguments{
\item{X}{Design matrix}

\item{y}{Response vector}

\item{num_split}{Repeated number of DS procedure for MDS}

\item{q}{FDR control level}
}
\value{
A list containing the selected variables using DS or MDS.
}
\description{
DS procedure first splits the data into two halves. In the first
half of the data, we apply LASSO for which the penalization parameter
\code{lambda} is chosen via 10-fold cross validation. In the second half
of the data, we apply OLS, and get the least squares estimator. We combine
the LASSO estimator and OLS estimatr via mirror statistics proposed in
\url{https://arxiv.org/abs/2002.08542}.
}
\examples{
require('glmnet')
require('MASS')
n = 500
p = 1000
p0 = 50
delta = 3
Sigma = diag(1, p)
data = generate_data(n, p, p0, Sigma, delta)
X = data$X
Y = data$y
signal_index = data$signal_index
num_split = 50
q = 0.1
selection = DS(X, Y, num_split, q)
DS_selected = selection$DS_feature
MDS_selected = selection$MDS_feature
DS_result = fdp_power(DS_selected, signal_index)
MDS_result = fdp_power(MDS_selected, signal_index)
c(DS_result$fdp, MDS_result$fdp)
c(DS_result$power, MDS_result$power)
}
