% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DS_graph.R
\name{DS_graph}
\alias{DS_graph}
\title{FDR control via DS and MDS in graphical model}
\usage{
DS_graph(data, q, num_split)
}
\arguments{
\item{data}{n by p matrix}

\item{q}{FDR control level}

\item{num_split}{Repeated number of DS procedure for MDS}
}
\value{
A list containing the selected edges using DS or MDS. Each component in the list is
a binary p by p matrix A, of which \code{A_{ij} = 1} indicates the edge between node \code{i}
and \code{j} is selected.
}
\description{
The estimation for gaussian graphical model can be recast as a nodewise
regression problem. Thus we apply our DS and MDS approach nodewisely and combine all
the nodewise selection results via \emph{or} rule. The nominal level for each nodewise
selection is set to be \code{q/2} in order to control the FDR for the whole graph properly.
}
\examples{
require('glmnet')
require('MASS')
n = 300; p = 50; rho = 8; a = -0.6; c = 1.5
q = 0.2
num_split = 50
precision = matrix(0, nrow = p, ncol = p)
edges_set = matrix(0, nrow = p, ncol = p)
###Construct banded graph
for(i in 1:p){
 for(j in 1:p){
   if(i == j){
     precision[i, j] <- 1
   }
   if(i != j & abs(i - j) <= rho){
     precision[i, j] <- sign(a)*abs(a)^(abs(i - j)/c)
     edges_set[i, j] <- 1
   }
  }
}
min_eigen = min(eigen(precision)$values)
if(min_eigen < 0){diag(precision) <- diag(precision) + abs(min_eigen) + 0.005}
### generate samples
data <- mvrnorm(n, mu = rep(0,p), Sigma = solve(precision))
selected = DS_graph(data, q, num_split)
DS_selected = selected$DS_selected_edge
MDS_selected = selected$MDS_selected_edge
DS_result = fdp_power_graph(DS_selected, edges_set)
MDS_result = fdp_power_graph(MDS_selected, edges_set)
c(DS_result$fdp, MDS_result$fdr)
c(DS_result$power, MDS_result$power)
}
