% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdp_power.R
\name{fdp_power}
\alias{fdp_power}
\title{FDP and power calculation}
\usage{
fdp_power(selected_index, signal_index)
}
\arguments{
\item{selected_index}{Index of selected features by DS or MDS}

\item{signal_index}{True relevent feature index.}
}
\value{
A list containing FDP and power.
}
\description{
Calculate FDP and power of DS and MDS method
}
