% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdp_power_graph.R
\name{fdp_power_graph}
\alias{fdp_power_graph}
\title{FDP and power calculation for gaussian graphical model.}
\usage{
fdp_power_graph(selected_edge, true_edge)
}
\arguments{
\item{selected_edge}{Binary matrix of which each element indicates whether the edge is
selected by DS or MDS.}

\item{true_edge}{Binary matrix of which each element indicates whether the edge exists.}
}
\value{
A list containing FDP and power.
}
\description{
FDP and power calculation for gaussian graphical model.
}
