% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Data generating process}
\usage{
generate_data(n, p, p0, covariance, delta)
}
\arguments{
\item{n}{Sample size}

\item{p}{Number of feature}

\item{p0}{Number of relevent features}

\item{covariance}{Covaraince matrix for the design matrix}

\item{delta}{Signal strength, the true signal strength is \deqn{delta*(log p/n)^{1/2}}}
}
\value{
A list containing design matrix, response variable and signal index
}
\description{
Generating data following linear regression
}
\examples{
require('MASS')
n = 500; p = 1000; p0 = 50; delta = 3
Sigma =  diag(1, p)
data = generate_data(n, p, p0, Sigma, delta)
X = data$X
Y = data$y
signal_index = data$signal_index
}
